% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{archipelenfeu}
\alias{archipelenfeu}
\title{L'archipel en feu}
\format{Un vecteur de 6282 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/cache/epub/17660/pg17660.txt}
}
\usage{
archipelenfeu
}
\description{
Le texte complet du livre "L'archipel en feu", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
