% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{autourdelalune}
\alias{autourdelalune}
\title{Autour de la lune}
\format{Un vecteur de 6024 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/files/46111/46111-0.txt}
}
\usage{
autourdelalune
}
\description{
Le texte complet du livre "Autour de la lune", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
