% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{cinqsemaines}
\alias{cinqsemaines}
\title{Cinq semaines en ballon}
\format{Un vecteur de 8688 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/cache/epub/4548/pg4548.txt}
}
\usage{
cinqsemaines
}
\description{
Le texte complet du livre "Cinq semaines en ballon", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
