% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{package}
\name{julesverne}
\alias{julesverne}
\alias{julesverne-package}
\title{Les livres de Jules Verne.}
\description{
Ce package contient quelques livres de Jules Verne, formattés pour
servir de jeux de données pour de l'analyse de texte.
}
