% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{souslesmers1}
\alias{souslesmers1}
\title{Vingt mille lieues sous les mers - première partie}
\format{Un vecteur de 6585 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/cache/epub/5095/pg5095.txt}
}
\usage{
souslesmers1
}
\description{
Le texte complet du livre "Vingt mille lieues sous les mers - première partie", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
