% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{souslesmers2}
\alias{souslesmers2}
\title{Vingt mille lieues sous les mers - deuxième partie}
\format{Un vecteur de 7750 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/cache/epub/5096/pg5096.txt}
}
\usage{
souslesmers2
}
\description{
Le texte complet du livre "Vingt mille lieues sous les mers - deuxième partie", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
