% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{tourdumonde}
\alias{tourdumonde}
\title{Le tour du monde en 80 jours}
\format{Un vecteur de 6989 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/cache/epub/800/pg800.txt}
}
\usage{
tourdumonde
}
\description{
Le texte complet du livre "Le tour du monde en 80 jours", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
