% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verne.R
\name{verne}
\alias{verne}
\title{Data frame de plusieurs livres de Jules Verne.}
\usage{
verne()
}
\value{
Un data frame de 2 colonnes: \code{text}: une ligne de texte, \code{book}:
le titre du livre correspondant.
}
\description{
Retourne un data frame de 2 colonnes: \code{text} qui contient le texte du livre
(environ 70 caractères par ligne), et \code{book} qui contient le titre
du livre correspondant. Les titres du livre sont de type \code{factor}.
}
\details{
Les textes sont issus du Projet Gutenberg (\url{http://www.gutenberg.org/})
et modifiés pour ne pas inclure les informations de license. Les lignes vides ont
également été retirées.
}
\examples{

library(dplyr)

verne() \%>\%
  group_by(book) \%>\%
  summarise(total_lines = n())

}
