% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{voyageterre}
\alias{voyageterre}
\title{Voyage au centre de la terre}
\format{Un vecteur de 7714 éléments de chaînes de caractères.}
\source{
\url{http://www.gutenberg.org/cache/epub/4791/pg4791.txt}
}
\usage{
voyageterre
}
\description{
Le texte complet du livre "Voyage au centre de la terre", encodé en UTF-8
tel que distribué sur le Projet Gutenberg.
}
\keyword{datasets}
